feat(#651): allow cpu-limitation
Added config to allow to limit the number of
cpu-threads to use for the worker-threads.
The new config is per default 0, which means
there is no limitation and all cpu-threads, expect
the 2 reserved ones, are allowed to be used as
worker-threads.
Also removed the old memory-limitation-config,
because there is no pre-allocation of memory anymore.

+++ src/binaries/hanami/src/core/processing/worker_thread.rs
@@ -26,7 +26,7 @@ use super::worker_queue::*;
 
 pub struct WorkerThread {
     #[allow(dead_code)]
-    pub thread_id: u64,
+    pub thread_id: usize,
 
     pub handle: Option<JoinHandle<()>>,
     pub running: Arc<AtomicBool>,
@@ -98,16 +98,14 @@ fn process_task(worker_task: &WorkerTask) -> Result<(), AinariError> {
 }
 
 impl WorkerThread {
-    pub fn new(thread_id: u64) -> Self {
+    pub fn new(thread_id: usize) -> Self {
         let running = Arc::new(AtomicBool::new(true));
         let running_clone = Arc::clone(&running);
         log::debug!("Create worker-thread on cpu-thread {thread_id}");
 
         let handle = thread::spawn(move || {
             log::debug!("Started worker-thread on cpu-thread {thread_id}");
-            let core_id = core_affinity::CoreId {
-                id: thread_id as usize,
-            };
+            let core_id = core_affinity::CoreId { id: thread_id };
             let res = core_affinity::set_for_current(core_id);
             if !res {
                 log::warn!("Failed to pin worker-thread to cpu-thread {thread_id}");
